Pygame.py has been made!!!
Finally got around to it! I was chipping away at it, but after I sat down for a couple hours to dedicate myself, it was really easy.
It has more features than last time. You can call for UNO, play cards, draw cards, and can click the arrows on the side (although the arrows are currently pretty janky due to having to wait)

There is an example at the bottom that gives a good example of how it works, but all you really need to do is make a PygameWrapper object with width and height, and then make a UserInterface object with that PygameWrapper, a discardPile, and a drawPile. 
Then, with a player object, in the loop you can get input from a player via userInput = UserInterface.interfaceUser(Player), which will either directly output a card they wish to play, a -1 to immediately close the game, 0 for saying UNO, and 1 for drawing a card. 
That also directly renders the game in the players eyes, although we don't currently see other players hands. I will get around to that in due time, alongside a main menu. But those aren't technically necessary for UNO to work, and we have more pressing matters!

Also, my code uses a different card class than the current, which is Ryan's card made here in this branch: https://github.com/jul1anpa/project-uno/tree/CardClass
No idea why he hasn't made a pull request; I reviewed and supported it. Maybe he just forgot?
In any case, it has issues I found making this that his branch hasn't fixed lol. So, I have an updated version of his card in the example code you can probably rip.