Added an assortment of new Pygame logic!
Particularly, the main menu, giving us four screens. It allows the user to start a new game and select X many players and A.I's, alongside windows options. You call menu.mainMenu() and it returns the amount of players the user wants in a game, with names and all!

Needs a lot of documentation, and not really the version I want to pull, so I've decided to make this a branch with a pull request so we can look at it during our meeting.